requirements.txt
requests
numpy
pygame
spotipy
Pillow
pyaudiowpatch
winsdk
pywin32
python-dotenv